Refactor provider Effect services
