+++ apps/server/src/provider/ProviderAdapterRegistry.test.ts
@@ -10,16 +10,15 @@ import * as Layer from "effect/Layer";
 import * as PubSub from "effect/PubSub";
 import * as Stream from "effect/Stream";
 
-import type * as ClaudeAdapter from "../Services/ClaudeAdapter.ts";
-import type * as CodexAdapter from "../Services/CodexAdapter.ts";
-import type * as CursorAdapter from "../Services/CursorAdapter.ts";
-import type * as OpenCodeAdapter from "../Services/OpenCodeAdapter.ts";
-import * as ProviderAdapterRegistry from "../Services/ProviderAdapterRegistry.ts";
-import * as ProviderInstanceRegistry from "../Services/ProviderInstanceRegistry.ts";
-import type { ProviderInstance } from "../ProviderDriver.ts";
-import { makeManualOnlyProviderMaintenanceCapabilities } from "../providerMaintenance.ts";
-import type * as TextGeneration from "../../textGeneration/TextGeneration.ts";
-import * as ProviderAdapterRegistryLayer from "./ProviderAdapterRegistry.ts";
+import type { ClaudeAdapterShape } from "./Services/ClaudeAdapter.ts";
+import type { CodexAdapterShape } from "./Services/CodexAdapter.ts";
+import type { CursorAdapterShape } from "./Services/CursorAdapter.ts";
+import type { OpenCodeAdapterShape } from "./Services/OpenCodeAdapter.ts";
+import * as ProviderAdapterRegistry from "./ProviderAdapterRegistry.ts";
+import * as ProviderInstanceRegistry from "./ProviderInstanceRegistry.ts";
+import type { ProviderInstance } from "./ProviderDriver.ts";
+import { makeManualOnlyProviderMaintenanceCapabilities } from "./providerMaintenance.ts";
+import type * as TextGeneration from "../textGeneration/TextGeneration.ts";
 import * as NodeServices from "@effect/platform-node/NodeServices";
 
 const CODEX_DRIVER = ProviderDriverKind.make("codex");
@@ -27,7 +26,7 @@ const CLAUDE_AGENT_DRIVER = ProviderDriverKind.make("claudeAgent");
 const OPENCODE_DRIVER = ProviderDriverKind.make("opencode");
 const CURSOR_DRIVER = ProviderDriverKind.make("cursor");
 
-const fakeCodexAdapter: CodexAdapter.CodexAdapterShape = {
+const fakeCodexAdapter: CodexAdapterShape = {
   provider: CODEX_DRIVER,
   capabilities: { sessionModelSwitch: "in-session" },
   startSession: vi.fn(),
@@ -44,7 +43,7 @@ const fakeCodexAdapter: CodexAdapter.CodexAdapterShape = {
   streamEvents: Stream.empty,
 };
 
-const fakeClaudeAdapter: ClaudeAdapter.ClaudeAdapterShape = {
+const fakeClaudeAdapter: ClaudeAdapterShape = {
   provider: CLAUDE_AGENT_DRIVER,
   capabilities: { sessionModelSwitch: "in-session" },
   startSession: vi.fn(),
@@ -61,7 +60,7 @@ const fakeClaudeAdapter: ClaudeAdapter.ClaudeAdapterShape = {
   streamEvents: Stream.empty,
 };
 
-const fakeOpenCodeAdapter: OpenCodeAdapter.OpenCodeAdapterShape = {
+const fakeOpenCodeAdapter: OpenCodeAdapterShape = {
   provider: OPENCODE_DRIVER,
   capabilities: { sessionModelSwitch: "in-session" },
   startSession: vi.fn(),
@@ -78,7 +77,7 @@ const fakeOpenCodeAdapter: OpenCodeAdapter.OpenCodeAdapterShape = {
   streamEvents: Stream.empty,
 };
 
-const fakeCursorAdapter: CursorAdapter.CursorAdapterShape = {
+const fakeCursorAdapter: CursorAdapterShape = {
   provider: CURSOR_DRIVER,
   capabilities: { sessionModelSwitch: "in-session" },
   startSession: vi.fn(),
@@ -95,7 +94,7 @@ const fakeCursorAdapter: CursorAdapter.CursorAdapterShape = {
   streamEvents: Stream.empty,
 };
 
-// ProviderAdapterRegistryLive is now a facade over ProviderInstanceRegistry —
+// ProviderAdapterRegistry is a facade over ProviderInstanceRegistry —
 // it walks `listInstances` once at boot and surfaces the default-instance
 // adapter keyed by its driver kind. To test the facade we supply four fake
 // instances whose `instanceId === defaultInstanceIdForDriver(driverKind)` so
@@ -147,14 +146,11 @@ const fakeInstanceRegistryLayer = Layer.succeed(ProviderInstanceRegistry.Provide
 });
 
 const layer = Layer.mergeAll(
-  Layer.provide(
-    ProviderAdapterRegistryLayer.ProviderAdapterRegistryLive,
-    fakeInstanceRegistryLayer,
-  ),
+  Layer.provide(ProviderAdapterRegistry.layer, fakeInstanceRegistryLayer),
   NodeServices.layer,
 );
 
-it.layer(layer)("ProviderAdapterRegistryLive", (it) => {
+it.layer(layer)("ProviderAdapterRegistry", (it) => {
   it("resolves adapters and routing metadata from provider instances", () =>
     Effect.gen(function* () {
       const registry = yield* ProviderAdapterRegistry.ProviderAdapterRegistry;
