+++ apps/server/src/provider/ProviderInstanceRegistry.ts
@@ -1,5 +1,5 @@
 /**
- * ProviderInstanceRegistryLive — runtime implementation of
+ * ProviderInstanceRegistry — runtime implementation of
  * `ProviderInstanceRegistry` plus its sibling mutator.
  *
  * Materializes every entry in a `ProviderInstanceConfigMap`:
@@ -30,7 +30,7 @@
  * tears every instance down in reverse order; closing a single instance
  * (via `reconcile` removing it) leaves the rest untouched.
  *
- * @module provider/Layers/ProviderInstanceRegistryLive
+ * @module provider/ProviderInstanceRegistry
  */
 import {
   defaultInstanceIdForDriver,
@@ -51,16 +51,69 @@ import * as Schema from "effect/Schema";
 import * as Scope from "effect/Scope";
 import * as Stream from "effect/Stream";
 
-import { buildUnavailableProviderSnapshot } from "../unavailableProviderSnapshot.ts";
-import {
+import { buildUnavailableProviderSnapshot } from "./unavailableProviderSnapshot.ts";
+import type { AnyProviderDriver, ProviderInstance } from "./ProviderDriver.ts";
+
+export class ProviderInstanceRegistry extends Context.Service<
   ProviderInstanceRegistry,
-  type ProviderInstanceRegistryShape,
-} from "../Services/ProviderInstanceRegistry.ts";
-import {
+  {
+    /**
+     * Look up one instance by id. Unknown ids return `undefined`; callers map
+     * that absence to the appropriate domain error.
+     */
+    readonly getInstance: (
+      instanceId: ProviderInstanceId,
+    ) => Effect.Effect<ProviderInstance | undefined>;
+
+    /**
+     * Every successfully materialized instance, in stable settings-author
+     * order.
+     */
+    readonly listInstances: Effect.Effect<ReadonlyArray<ProviderInstance>>;
+
+    /**
+     * Shadow snapshots for unknown drivers or invalid configurations, ready
+     * to merge into `ProviderRegistry` output.
+     */
+    readonly listUnavailable: Effect.Effect<ReadonlyArray<ServerProvider>>;
+
+    /**
+     * Emits after the registry adds, removes, or rebuilds instances. The
+     * payload is `void` because consumers re-read both registry lists.
+     *
+     * `Stream.fromPubSub` subscribes only when execution begins, so a newly
+     * forked consumer can miss a publish. Consumers that cannot tolerate that
+     * gap should acquire `subscribeChanges` first.
+     */
+    readonly streamChanges: Stream.Stream<void>;
+
+    /**
+     * Acquire a scoped PubSub subscription synchronously before forking the
+     * consumer loop, ensuring subsequent publishes cannot land in a gap.
+     */
+    readonly subscribeChanges: Effect.Effect<PubSub.Subscription<void>, never, Scope.Scope>;
+  }
+>()("t3/provider/ProviderInstanceRegistry") {}
+
+/**
+ * Internal mutation boundary used only by registry hydration.
+ *
+ * `reconcile` diffs by instance id, closes removed or replaced child scopes
+ * before creating replacements, materializes additions in fresh child scopes,
+ * and publishes one change tick after the batch. Reapplying an unchanged map
+ * is a no-op.
+ */
+export class ProviderInstanceRegistryMutator extends Context.Service<
   ProviderInstanceRegistryMutator,
-  type ProviderInstanceRegistryMutatorShape,
-} from "../Services/ProviderInstanceRegistryMutator.ts";
-import type { AnyProviderDriver, ProviderInstance } from "../ProviderDriver.ts";
+  {
+    /**
+     * Reconcile the live registry with a new configuration map. Individual
+     * driver creation failures become unavailable shadow snapshots so the
+     * settings watcher itself never fails.
+     */
+    readonly reconcile: (configMap: ProviderInstanceConfigMap) => Effect.Effect<void>;
+  }
+>()("t3/provider/ProviderInstanceRegistry/ProviderInstanceRegistryMutator") {}
 
 /**
  * Live registry entry: the materialized `ProviderInstance` + the fresh
@@ -315,9 +368,9 @@ const makeReconcile = <R>(input: {
  * Build the registry's runtime state from a concrete configMap. Returns a
  * record containing:
  *
- *   - `registry`: the read-only `ProviderInstanceRegistryShape` to expose
+ *   - `registry`: the read-only `ProviderInstanceRegistry["Service"]` to expose
  *     under `ProviderInstanceRegistry`.
- *   - `mutator`: the `ProviderInstanceRegistryMutatorShape` to expose
+ *   - `mutator`: the `ProviderInstanceRegistryMutator["Service"]` to expose
  *     under `ProviderInstanceRegistryMutator`.
  *   - `reconcile`: the raw reconcile function, provided for convenience so
  *     boot-time layers can hydrate an initial map before publishing the
@@ -327,13 +380,13 @@ const makeReconcile = <R>(input: {
  * created during `reconcile`. Closing that scope closes every live
  * instance.
  */
-export const makeProviderInstanceRegistry = <R>(input: {
+export const make = <R>(input: {
   readonly drivers: ReadonlyArray<AnyProviderDriver<R>>;
   readonly configMap: ProviderInstanceConfigMap;
 }): Effect.Effect<
   {
-    readonly registry: ProviderInstanceRegistryShape;
-    readonly mutator: ProviderInstanceRegistryMutatorShape;
+    readonly registry: ProviderInstanceRegistry["Service"];
+    readonly mutator: ProviderInstanceRegistryMutator["Service"];
   },
   never,
   R | Scope.Scope
@@ -362,14 +415,14 @@ export const makeProviderInstanceRegistry = <R>(input: {
 
     const state: RegistryState = { entries, unavailable, changes };
     const reconcileWithR = makeReconcile({ state, driversById, parentScope });
-    const reconcile: ProviderInstanceRegistryMutatorShape["reconcile"] = (configMap) =>
+    const reconcile: ProviderInstanceRegistryMutator["Service"]["reconcile"] = (configMap) =>
       reconcileWithR(configMap).pipe(Effect.provideContext(driverContext));
 
     // Hydrate the initial configMap synchronously so callers can read
     // `listInstances` immediately after this effect completes.
     yield* reconcile(input.configMap);
 
-    const registry: ProviderInstanceRegistryShape = {
+    const registry = ProviderInstanceRegistry.of({
       getInstance: (id) => Ref.get(entries).pipe(Effect.map((map) => map.get(id)?.instance)),
       listInstances: Ref.get(entries).pipe(
         Effect.map(
@@ -395,9 +448,9 @@ export const makeProviderInstanceRegistry = <R>(input: {
       get subscribeChanges() {
         return PubSub.subscribe(changes);
       },
-    };
+    });
 
-    const mutator: ProviderInstanceRegistryMutatorShape = { reconcile };
+    const mutator = ProviderInstanceRegistryMutator.of({ reconcile });
 
     return { registry, mutator };
   });
@@ -409,15 +462,15 @@ export const makeProviderInstanceRegistry = <R>(input: {
  * wiring up the settings watcher.
  *
  * Only exposes the public registry tag — hot-reload consumers should use
- * `ProviderInstanceRegistryMutableLayer` (below) or the hydration layer.
+ * `mutableLayer` (below) or the hydration layer.
  */
-export const ProviderInstanceRegistryLayer = <R>(input: {
+export const layer = <R>(input: {
   readonly drivers: ReadonlyArray<AnyProviderDriver<R>>;
   readonly configMap: ProviderInstanceConfigMap;
 }): Layer.Layer<ProviderInstanceRegistry, never, R> =>
   Layer.effect(
     ProviderInstanceRegistry,
-    makeProviderInstanceRegistry(input).pipe(Effect.map((built) => built.registry)),
+    make(input).pipe(Effect.map((built) => built.registry)),
   ) as Layer.Layer<ProviderInstanceRegistry, never, R>;
 
 /**
@@ -426,12 +479,12 @@ export const ProviderInstanceRegistryLayer = <R>(input: {
  * changes. Tests that exercise the mutator directly can pair this Layer
  * with a test-local `ServerSettingsService`.
  */
-export const ProviderInstanceRegistryMutableLayer = <R>(input: {
+export const mutableLayer = <R>(input: {
   readonly drivers: ReadonlyArray<AnyProviderDriver<R>>;
   readonly configMap: ProviderInstanceConfigMap;
 }): Layer.Layer<ProviderInstanceRegistry | ProviderInstanceRegistryMutator, never, R> =>
   Layer.effectContext(
-    makeProviderInstanceRegistry(input).pipe(
+    make(input).pipe(
       Effect.map(({ registry, mutator }) =>
         Context.make(ProviderInstanceRegistry, registry).pipe(
           Context.add(ProviderInstanceRegistryMutator, mutator),
