+++ apps/server/src/provider/testUtils/providerRegistryMock.ts
@@ -1,4 +1,4 @@
-import { ProviderRegistry, type ProviderRegistryShape } from "../Services/ProviderRegistry.ts";
+import * as ProviderRegistry from "../ProviderRegistry.ts";
 import type { ServerProvider } from "@t3tools/contracts";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
@@ -7,7 +7,7 @@ import { makeManualOnlyProviderMaintenanceCapabilities } from "../providerMainte
 
 export const makeProviderRegistryMock = (
   providers: ReadonlyArray<ServerProvider> = [],
-): ProviderRegistryShape => ({
+): ProviderRegistry.ProviderRegistry["Service"] => ({
   getProviders: Effect.succeed(providers),
   refresh: () => Effect.succeed(providers),
   refreshInstance: () => Effect.succeed(providers),
@@ -18,4 +18,4 @@ export const makeProviderRegistryMock = (
 });
 
 export const makeProviderRegistryLayer = (providers: ReadonlyArray<ServerProvider> = []) =>
-  Layer.succeed(ProviderRegistry, makeProviderRegistryMock(providers));
+  Layer.succeed(ProviderRegistry.ProviderRegistry, makeProviderRegistryMock(providers));
