+++ oxlint-plugin-t3code/rules/no-manual-effect-runtime-in-tests.ts
@@ -40,8 +40,8 @@ const LEGACY_BASELINE = new Map<string, number>([
   ["apps/server/src/provider/Layers/CodexSessionRuntime.test.ts", 5],
   ["apps/server/src/provider/Layers/CursorAdapter.test.ts", 1],
   ["apps/server/src/provider/Layers/CursorProvider.test.ts", 4],
-  ["apps/server/src/provider/Layers/ProviderService.test.ts", 2],
-  ["apps/server/src/provider/Layers/ProviderSessionReaper.test.ts", 21],
+  ["apps/server/src/provider/ProviderService.test.ts", 2],
+  ["apps/server/src/provider/ProviderSessionReaper.test.ts", 21],
   ["apps/server/src/relay/AgentAwarenessRelay.test.ts", 4],
   ["apps/server/src/server.test.ts", 1],
   ["apps/web/src/cloud/dpop.test.ts", 2],
